Machine Learning/requirements.txt
numpy
faiss-cpu
sentence-transformers
torch
pymongo
python-dotenv
fastapi
uvicorn
pydantic


